Support GET_TS by simple memcpy
It is a naive approach. We could meassure the performance and use it as
a baseline value. Some optimized implementation should get guage values
in parallel, and cache it on GPU's memory.

Besides, calculating the location of a guage on every time step is a
waste of time on host side. And memory copy should be async as well.

========================================================

Here is a profiler result for this implementation on 40 guaged
locations:
            Type  Time(%)      Time     Calls       Avg       Min       Max  Name
 GPU activities:   31.13%  42.8077s     43200  990.92us  963.04us  1.0469ms  momt_kernelN(GPU_Layer)
                   29.52%  40.6028s     43200  939.88us  902.44us  1.0373ms  momt_kernelM(GPU_Layer)
                   26.70%  36.7272s     43200  850.17us  831.81us  883.72us  mass_kernel(GPU_Layer)
                    9.89%  13.6035s     43200  314.90us  283.75us  370.40us  maximum_recorder_kernel(float*, float const *, unsigned int)
                    2.25%  3.09185s   3456250     894ns     640ns  4.1342ms  [CUDA memcpy DtoH]
                    0.50%  681.30ms    172800  3.9420us  1.9520us  19.680us  openbd_kernel(GPU_Layer, bd_side)
                    0.01%  15.921ms         9  1.7690ms     672ns  3.4719ms  [CUDA memcpy HtoD]
                    0.00%  2.5920us         4     648ns     608ns     768ns  [CUDA memset]
                    0.00%  1.2800us         1  1.2800us  1.2800us  1.2800us  [CUDA memcpy DtoD]

We can see that memcpy DtoH only contributes 9.89% to computation time.
Though it was increased by 7.481 times comparing to one without GET_TS
functionality, optimizing peice of code here may have little overall
improvement in common usecases of number of guages (n=40).

Fix: Correct indexing policy at GPUOutput

Fortran has array indices starting from 1 but it's 0 for C.